del rc file and other minor changes
