+++ ganga/GangaCore/test/gangaDB/db_controllers/temp_test.py
@@ -1,148 +1,148 @@
-"""Test the singularity container
-Custom test for singularity container: not used here
-"""
+# """Test the singularity container
+# Custom test for singularity container: not used here
+# """
 
-import os
-import time
-import json
-import utils
-import pytest
-from GangaCore.Core.GangaRepository.container_controllers import get_database_config
-from GangaCore.Utility.Config import getConfig
-from GangaCore.Utility.Virtualization import (
-    checkNative, checkDocker,
-    checkUDocker, checkSingularity
-)
-from GangaCore.testlib.GangaUnitTest import GangaUnitTest
-from GangaCore.Core.GangaRepository.container_controllers import singularity_handler, mongod_exists, ContainerCommandError
+# import os
+# import time
+# import json
+# import utils
+# import pytest
+# from GangaCore.Core.GangaRepository.container_controllers import get_database_config
+# from GangaCore.Utility.Config import getConfig
+# from GangaCore.Utility.Virtualization import (
+#     checkDocker,
+#     checkUDocker, checkSingularity
+# )
+# from GangaCore.testlib.GangaUnitTest import GangaUnitTest
+# from GangaCore.Core.GangaRepository.container_controllers import singularity_handler, mongod_exists, ContainerCommandError, checkNative
 
-HOST, PORT = utils.get_host_port()
-
-
-class SingularityTestHandler(GangaUnitTest):
-    """
-    Singularity Backend Controllers
-    """
-
-    def setUp(self):
-        """
-        """
-        extra_opts = utils.get_options(HOST, PORT)
-        extra_opts.append(("TestingFramework", "Flag", True))
-        self.sing_installated = checkSingularity()
-        database_config = get_database_config(None)
-        super(SingularityTestHandler, self).setUp(extra_opts=extra_opts)
-
-    def test_a1_setup_sif_file(self):
-        """
-        Download the sif file requried
-        """
-        if os.path.isfile(os.path.join(self.gangadir(), "mongo.sif")):
-            import gdown  # download the required sif file
-            url = 'https://drive.google.com/uc?id=1Z7k9LoFxGQKMjLzoe1D_jL9m5ReeixXk'
-            output = 'mongo.sif'
-            gdown.download(url, output, quiet=True)
-
-            import shutil  # copy the files from repository to testing directory
-            shutil.copy(
-                src="mongo.sif",
-                dst=self.gangadir()
-            )
-
-        # create logs folder in the testing dir
-        os.makedirs(os.path.join(self.gangadir(), "logs"), exist_ok=True)
-        assert True
+# HOST, PORT = utils.get_host_port()
 
 
-    def test_a2_singularity_handlers(self):
-        """
-        Test the starting and shutdown of udocker container image
-        """
-        # try:
-        # start the singularity container
-        singularity_handler(
-            action="start",
-            gangadir=self.gangadir(),
-            database_config=database_config
-        )
+# class SingularityTestHandler(GangaUnitTest):
+#     """
+#     Singularity Backend Controllers
+#     """
 
-        # checking if the container started up
-        flag = mongod_exists(
-            controller="singularity", cname=os.path.join(self.gangadir(), "mongo.sif")
-        )
-        assert flag is not None
+#     def setUp(self):
+#         """
+#         """
+#         extra_opts = utils.get_options(HOST, PORT)
+#         extra_opts.append(("TestingFramework", "Flag", True))
+#         self.sing_installated = checkSingularity()
+#         database_config = get_database_config(None)
+#         super(SingularityTestHandler, self).setUp(extra_opts=extra_opts)
 
-            # shutting down the container
-            singularity_handler(
-                action="quit",
-                gangadir=self.gangadir(),
-                database_config=database_config
-            )
-            flag = mongod_exists(
-                controller="singularity", cname=os.path.join(self.gangadir(), "mongo.sif")
-            )
-            assert flag is None
-        except Exception as e:
-            print(open(os.path.join(
-                self.gangadir(), "logs", "mongod-ganga.log"
-            )).read())
-            raise e
+#     def test_a1_setup_sif_file(self):
+#         """
+#         Download the sif file requried
+#         """
+#         if os.path.isfile(os.path.join(self.gangadir(), "mongo.sif")):
+#             import gdown  # download the required sif file
+#             url = 'https://drive.google.com/uc?id=1Z7k9LoFxGQKMjLzoe1D_jL9m5ReeixXk'
+#             output = 'mongo.sif'
+#             gdown.download(url, output, quiet=True)
 
-    def test_b_singularity_check_logs_created(self):
-        """
-        Check if the logs were successfully created
-        """
+#             import shutil  # copy the files from repository to testing directory
+#             shutil.copy(
+#                 src="mongo.sif",
+#                 dst=self.gangadir()
+#             )
 
-        log_path = os.path.join(
-            self.gangadir(), "logs", "mongod-ganga.log"
-        )
+#         # create logs folder in the testing dir
+#         os.makedirs(os.path.join(self.gangadir(), "logs"), exist_ok=True)
+#         assert True
 
-        assert os.path.exists(log_path)
 
-    def test_c_check_notimplemented_actions(self):
-        """
-        test not implemented error is raised
-        """
-        with pytest.raises(NotImplementedError) as _:
-            random_names = ["starts", "quite"]
-            for action in random_names:
-                singularity_handler(
-                    action=action,
-                    gangadir=self.gangadir(),
-                    database_config=database_config
-                )
+#     def test_a2_singularity_handlers(self):
+#         """
+#         Test the starting and shutdown of udocker container image
+#         """
+#         # try:
+#         # start the singularity container
+#         singularity_handler(
+#             action="start",
+#             gangadir=self.gangadir(),
+#             database_config=database_config
+#         )
 
-    def test_d_permission_error_singularity_handler(self):
-        """
-        Remove the data dir from $GANGADIR
-        and check the expected behaviour from the logs
-        """
-        # removing the folder
-        data_path = os.path.join(self.gangadir(), "data")
+#         # checking if the container started up
+#         flag = mongod_exists(
+#             controller="singularity", cname=os.path.join(self.gangadir(), "mongo.sif")
+#         )
+#         assert flag is not None
 
-        os.system(f"chmod a-x {data_path}")
+#             # shutting down the container
+#             singularity_handler(
+#                 action="quit",
+#                 gangadir=self.gangadir(),
+#                 database_config=database_config
+#             )
+#             flag = mongod_exists(
+#                 controller="singularity", cname=os.path.join(self.gangadir(), "mongo.sif")
+#             )
+#             assert flag is None
+#         except Exception as e:
+#             print(open(os.path.join(
+#                 self.gangadir(), "logs", "mongod-ganga.log"
+#             )).read())
+#             raise e
 
-        with pytest.raises(ContainerCommandError) as _:
-            # start the handler, expecting an error
-            singularity_handler(
-                action="start",
-                gangadir=self.gangadir(),
-                database_config=database_config
-            )
+#     def test_b_singularity_check_logs_created(self):
+#         """
+#         Check if the logs were successfully created
+#         """
 
-        flag = mongod_exists(
-            controller="singularity", cname=os.path.join(self.gangadir(), "mongo.sif")
-        )
-        assert flag is None
+#         log_path = os.path.join(
+#             self.gangadir(), "logs", "mongod-ganga.log"
+#         )
 
-        # check the logs
-        log_path = os.path.join(
-            self.gangadir(), "logs", "mongod-ganga.log"
-        )
-        err_string = open(log_path, "r").read()
-        for _log in err_string.split("\n"):
-            if _log:
-                log = json.loads(_log)
-                if log['s'] == "E":
-                    assert "Unable to determine status of lock file in the data directory" in log[
-                        'attr']['error']
+#         assert os.path.exists(log_path)
+
+#     def test_c_check_notimplemented_actions(self):
+#         """
+#         test not implemented error is raised
+#         """
+#         with pytest.raises(NotImplementedError) as _:
+#             random_names = ["starts", "quite"]
+#             for action in random_names:
+#                 singularity_handler(
+#                     action=action,
+#                     gangadir=self.gangadir(),
+#                     database_config=database_config
+#                 )
+
+#     def test_d_permission_error_singularity_handler(self):
+#         """
+#         Remove the data dir from $GANGADIR
+#         and check the expected behaviour from the logs
+#         """
+#         # removing the folder
+#         data_path = os.path.join(self.gangadir(), "data")
+
+#         os.system(f"chmod a-x {data_path}")
+
+#         with pytest.raises(ContainerCommandError) as _:
+#             # start the handler, expecting an error
+#             singularity_handler(
+#                 action="start",
+#                 gangadir=self.gangadir(),
+#                 database_config=database_config
+#             )
+
+#         flag = mongod_exists(
+#             controller="singularity", cname=os.path.join(self.gangadir(), "mongo.sif")
+#         )
+#         assert flag is None
+
+#         # check the logs
+#         log_path = os.path.join(
+#             self.gangadir(), "logs", "mongod-ganga.log"
+#         )
+#         err_string = open(log_path, "r").read()
+#         for _log in err_string.split("\n"):
+#             if _log:
+#                 log = json.loads(_log)
+#                 if log['s'] == "E":
+#                     assert "Unable to determine status of lock file in the data directory" in log[
+#                         'attr']['error']
