+++ ganga/GangaCore/test/gangaDB/db_testing/TestGangaDBNested.py
@@ -4,7 +4,7 @@ import pymongo
 
 from GangaCore.Utility.Config import getConfig
 from GangaCore.testlib.GangaUnitTest import GangaUnitTest
-from GangaCore.Utility.Virtualization import checkNative, checkDocker
+from GangaCore.Utility.Virtualization import checkDocker
 
 
 HOST, PORT = utils.get_host_port()
